Fix rounding-error madness.

+++ opencog/atomspace/AttentionBank.h
@@ -230,7 +230,8 @@ public:
     /** Change the Very-Long-Term Importance of an attention value holder */
     //void setVLTI(AttentionValueHolderPtr avh, AttentionValue::vlti_t);
 
-    /** Retrieve the doubly normalised Short-Term Importance between -1..1
+    /**
+     * Retrieve the doubly normalised Short-Term Importance between -1..1
      * for a given AttentionValue. STI above and below threshold
      * normalised separately and linearly.
      *
@@ -241,13 +242,15 @@ public:
      *        Outside this range can be return if average=true
      * @return normalised STI between -1..1
      */
-    float getNormalisedSTI(AttentionValuePtr, bool average, bool clip) const;
+    double getNormalisedSTI(AttentionValuePtr, bool average, bool clip) const;
 
     /**
      * @see getNormalisedSTI()
      */
-    float getNormalisedSTI(AttentionValuePtr) const;
-    /** Retrieve the linearly normalised Short-Term Importance between 0..1
+    double getNormalisedSTI(AttentionValuePtr) const;
+
+    /**
+     * Retrieve the linearly normalised Short-Term Importance between 0..1
      * for a given AttentionValue.
      *
      * @param h The attention value holder to get STI for
@@ -257,7 +260,7 @@ public:
      *        Outside this range can be return if average=true
      * @return normalised STI between 0..1
      */
-    float getNormalisedZeroToOneSTI(AttentionValuePtr, bool average, bool clip) const;
+    double getNormalisedZeroToOneSTI(AttentionValuePtr, bool average, bool clip) const;
 };
 
 /** @}*/
